Unwind panics: Ignore when halted, abort if any other
