+++ Cargo.toml
@@ -231,6 +231,7 @@ version = "3.3"
 version = "2"
 
 [profile.release]
+panic = "unwind"
 opt-level = 3
 lto = "thin"
 incremental = true
