+++ console/network/environment/src/environment.rs
@@ -72,8 +72,7 @@ pub trait Environment:
     /// Halts the program from further synthesis, evaluation, and execution in the current environment.
     fn halt<S: Into<String>, T>(message: S) -> T {
         println!("{}", message.into());
-        exit(0);
-        unreachable!()
+        panic!("HaltedABC")
     }
 }
